Changed length parameters from uint8_t to uint16_t in LocalUDP and EthernetTCP strategies.

+++ strategies/EthernetTCP/EthernetTCP.h
@@ -33,10 +33,10 @@ class EthernetTCP {
     uint8_t incoming_packet_buf[PACKET_MAX_LENGTH];
     uint16_t incoming_packet_size = 0;
     uint16_t incoming_packet_pos = 0;
-    static void static_receiver(uint8_t id, const uint8_t *payload, uint8_t length, void *callback_object) {
+    static void static_receiver(uint8_t id, const uint8_t *payload, uint16_t length, void *callback_object) {
       if (callback_object) ((EthernetTCP*)callback_object)->receiver(id, payload, length);
     }
-    void receiver(uint8_t id, const uint8_t *payload, uint8_t length) {
+    void receiver(uint8_t id, const uint8_t *payload, uint16_t length) {
       if (length <= PACKET_MAX_LENGTH) {
         memcpy(incoming_packet_buf, payload, length);
         incoming_packet_size = length;
@@ -82,7 +82,7 @@ class EthernetTCP {
 
     /* Send a string: */
 
-    void send_string(uint8_t *string, uint8_t length) {
+    void send_string(uint8_t *string, uint16_t length) {
       if (length > 0)
         last_send_result = link.send((uint8_t)string[0], (const char*)string, length);
     };
